delete a couple of unused patches
* delete the disabled 60_add-ant-hack.patch

Upstream removed ant because of name clash, we let the patch
stay (but unused) in case some downstream liked it. Rather
wait for it to be renamed, then we might put it back.

* delete the unused 52_driver_XScreenSaver.ad.in.patch

It was left over for reference in case we had missed
something from it, but it's time to clean it away.

If needed we can find it in the git history instead.