delete popsquares.desktop so it won't be shipped
This file was not included in xscreensaver-data.files so we would
ship the hack but not the desktop file because of the unresolved
conflict with the screensaver and desktop file of the same name
in gnome-screensaver.

With the new split-hacks.sh, any existing desktop file would have been
shipped together with the hack.

We will have to keep this is mind when we move to autogeneration
on the fly of the desktop files.