Only delete configuration files if we possibly shipped them
otherwise we might delete files which the user explicitly copied
back (having the same contents).

For now we only check if we are upgrading from a version from
before we started deleting those files unconditionally. To be
more careful we could also check that we are not upgrading from
a version _before_ we started shipping those files. That would
cover the corner case of someone having installed their own
files before we started shipping them, and those files matching
exactly ours (which is not so likely).