Build without gdm and remove obsolete gdmflexiserver options
We were already prepared for this:

The "open url" and "open file" commands are patched to not depend on the
build environment. The commands are gnome-agnostic.

Also, instead of having configure detect gdm on the build machine to
determine whether gdmflexiserver should be used as newLoginCommand we
have been hardcoding it anyway.

An old, remaining bug is that the "New login" is displayed on machines
which do not have the gdmflexiserver installed, unless the
newLoginCommand entry in app-defaults is commented out.

A run-time check would be needed in the xscreensaver code, to verify
that the newLoginCommand binary actually exists on the machine. TODO.

Also, a wrapper could be needed to call either gdmflexiserver or kdmctl
depending on which login server is running on the machine. TODO.