Build with -D_FILE_OFFSET_BITS=64 to fix inode number overflow
This flag should be added properly in configure.in upstream.
Since we are not rebuilding configure, I have just added this
to debian/rules for now.

Analysis by Mamoru Tasaka:
"It seems that stat() on the line 1256 in get_filename_1() in
xscreensaver-getimage.c returns -1 with errno 75 = EOVERFLOW. stat(2)
says that this can happen when such application was not compiled with
-D_FILE_OFFSET_BITS=64 and when using 32-bit platform. However while
stat(2) says that this can happen when accessing too large file, this
issue is not this case. For this case, EOVERFLOW happens because files
on cifs-mounted system can return very large i-node : for people seeing
this issue you can check this by "ls -ali" on cifs-mounted system."

https://bugs.launchpad.net/bugs/609451