Add a Gnome/KDE-aware xscreensaver daemon wrapper
The wrapper checks if we are running a Gnome session and
if gdmflexiserver can be used for the New Login button.

A similar check is done for KDE / kdmctl.

Modify the daemon example desktop file to use the wrapper.
Also add X-KDE-StartupNotify=false for more KDE friendliness.

The master plan is to have the environment which starts the xscreensaver
daemon specify how new sessions can be launched, and to take any
second-guessing and references to gdmflexiserver and friends out of
xscreensaver. We will then be able to use a run-time check of the
corresponding X resource to determine whether the "New login" button
should be displayed or not.

See http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=304406