Add patch to speed up locking from suspend scripts
Some suspend infrastructures call xscreensaver-command --lock
but puts the computer to sleep before the screen is actually
locked. screensaver-command always waited a second before
activating the saver to avoid waking up from the key release
when the command is typed in, and this made the race condition
even worse.

This patch skips this wait if the command is not run from a tty.

https://bugs.launchpad.net/bugs/1229486

It does not solve the problem of racy suspend scripts though. They
should wait for the completion of the xscreensaver-command:

http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=593303